requirements.txt
streamlit
pymupdf
pytesseract
pillow
google-generativeai
fpdf2